requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.9.1
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
cbor==1.0.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
pysqlite3-binary==0.5.4
chromadb==1.0.9
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
datasets==4.0.0
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
durationpy==0.10
elastic-transport==9.1.0
elasticsearch==9.1.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.9
fastjsonschema==2.21.1
filelock==3.18.0
FlagEmbedding==1.3.5
flatbuffers==25.2.10
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-genai==1.27.0
googleapis-common-protos==1.70.0
gotrue==2.12.3
grpcio==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
ijson==3.4.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
inscriptis==2.6.0
ipykernel==6.30.0
ipython==8.37.0
ir_datasets==0.5.11
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kubernetes==33.1.0
lark==1.2.2
lxml==6.0.0
lz4==4.4.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mmh3==5.1.0
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
narwhals==2.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.4.4
notebook_shim==0.2.4
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.23.0
groq==0.31.0
openpyxl==3.1.5
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-instrumentation==0.56b0
opentelemetry-instrumentation-asgi==0.56b0
opentelemetry-instrumentation-fastapi==0.56b0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
opentelemetry-util-http==0.56b0
orjson==3.11.1
optimum==1.23.3
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
peft==0.16.0
pillow==11.3.0
platformdirs==4.3.8
plotly==6.3.0
portalocker==3.2.0
postgrest==1.0.2
posthog==6.3.1
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.13.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
python-json-logger==3.3.0
pytz==2025.2
pywin32==311; sys_platform == "win32"
pywinpty==2.0.15; sys_platform == "win32"
PyYAML==6.0.2
pyzmq==27.0.0
qdrant-client==1.15.0
realtime==2.4.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rouge==1.0.1
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
Send2Trash==1.8.3
sentence-transformers==5.0.0
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.45.3
storage3==0.11.3
streamlit==1.47.1
StrEnum==0.4.15
supafunc==0.9.4
sympy==1.14.0
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.0
trec-car-tools==2.6
typer==0.16.0
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
unlzw3==0.2.3
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.35.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zlib-state==0.1.9
